\name{getBaseDirectory}
\alias{getBaseDirectory}
\title{Select the base directory for the NAEP data}
\usage{
  getBaseDirectory()
}
\value{
  string with the path to the NAEP data
}
\description{
  Select the root directory of NAEP data disc.
}
\details{
  Provides a directory selection box to select the root of
  the NAEP data DVD. This function also provides some basic
  verification that the directory selected is a proper NAEP
  dataset.
}

