\docType{package}
\name{naep-package}
\alias{naep}
\alias{naep-package}
\title{National Assessment of Educational Progress (NAEP)}
\description{
  National Assessment of Educational Progress (NAEP)
}
\details{
  This package provides functions to read data from NAEP
  restricted data CDs or DVDs. This obviates the need for
  the NAEP Data Explorer (which is Windows only). The
  resulting data frame will have correctly coded varaibles.
  Additional functions provide access to the data
  dictionary as a data frame. Variable selection can be
  done through a dialog box or by character lists.
}
\author{
  \email{jason@bryer.org}
}
\keyword{data}
\keyword{naep}
\keyword{nces}
\keyword{package}

