\name{print.naep.model}
\alias{print.naep.model}
\title{Prints summary of the combined NAEP model.}
\usage{
  print.naep.model(x, ...)
}
\arguments{
  \item{x}{a naep.model object.}

  \item{...}{other parameters passed to \code{print}.}
}
\description{
  Prints summary of the combined NAEP model.
}

