\name{summary.naep.model}
\alias{summary.naep.model}
\title{Prints summary of the combined NAEP model.}
\usage{
  summary.naep.model(object, ...)
}
\arguments{
  \item{object}{a naep.model object.}

  \item{...}{other parameters passed to \code{summary}.}
}
\description{
  Prints summary of the combined NAEP model.
}

